% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_example.R
\name{rbivariate.rwmh}
\alias{rbivariate.rwmh}
\title{A random walk Metropolis–Hastings algorithm for Generating the
bivariate Gaussian distribution.}
\usage{
rbivariate.rwmh(N, burn, Y0, sigma_matrix, mu_vec)
}
\arguments{
\item{N}{represents the number of samples to retain after discarding the
number of burn-in samples.}

\item{burn}{of initial samples to discard as burn-in, to reduce the
dependence from the starting state.}

\item{Y0}{represents the initial state of the MCMC chain.}

\item{sigma_matrix}{a 2x2 positive definite covariance matrix of the
target bivariate normal distribution from which we want to sample.}

\item{mu_vec}{a vector of size 2 that represents the means of the
target bivariate normal distribution from which we want to sample.}
}
\value{
an Nx2 matrix of samples generated using the random walk
Metropolis–Hastings algorithm.
}
\description{
Here, the proposal distribution is a random
walk with independent normal increments whose variances are taken from the
diagonal of the 2x2 positive definite covariance matrix.
}
\examples{
set.seed(1)
N <- 100000
burn <- 100
rho <- 0.7
sigma_matrix <- matrix(c(1, rho, rho, 1), nrow = 2, ncol = 2)
mu_vec <- c(0, 0)
test_pt1 <- rmvnorm(n = 1, mean = c(0, 0), sigma = sigma_matrix)
sim_attempt2 <- rbivariate.rwmh(N, burn, test_pt1, sigma_matrix, mu_vec)
}
